physics: force refactor
