+++ Cargo.toml
@@ -23,21 +23,25 @@ dev_native = [
     # Enable system information plugin for native dev builds.
     "bevy/sysinfo_plugin",
 ]
-configs = ["ron", "bevy_common_assets"]
+config-files = ["ron", "bevy_common_assets", "serde"]
 inspect = ["bevy-inspector-egui"]
 
 [dependencies]
-# required dependencies
-avian3d = { version = "0.1.2", features = ["debug-plugin"] }
+# core dependencies
 bevy = "0.14.2"
 bevy-trait-query = "0.6.0"
-iyes_perf_ui = { version = "0.3.0" }
 rand = "0.8.5"
-serde = { version = "1.0", features = ["derive"] }
-# optional dependencies
-ron = { version = "0.8.1", optional = true }
-bevy_common_assets = { version = "0.11.0", features = ["ron"], optional = true }
+# physics dependencies
+avian3d = { version = "0.1.2", features = ["debug-plugin"] }
+rayon = "1.10.0"
+# ui dependencies
 bevy-inspector-egui = { version = "0.27.0", features = ["highlight_changes"], optional = true }
+bevy_prototype_debug_lines = { version = "0.12.0", features = ["3d"] }
+iyes_perf_ui = { version = "0.3.0" }
+# file io dependencies
+bevy_common_assets = { version = "0.11.0", features = ["ron"], optional = true }
+ron = { version = "0.8.1", optional = true }
+serde = { version = "1.0", features = ["derive"], optional = true }
 
 [[bin]]
 name = "yahs"
