+++ src/simulator/mod.rs
@@ -1,4 +1,5 @@
-pub mod aero;
+#![allow(unused_imports)]
+
 pub mod atmosphere;
 pub mod balloon;
 pub mod forces;
